# api/openapi.yaml
openapi: 3.0.3
info:
  title: "RiPPository API"
  version: "3.0"
  description: "The RiPPository API allows retrieval and filtering of RiPP records by chemical or experimental parameters, similarity search on peak lists, and access to individual record details, metadata, service status, and available filter options."
paths:
...
  /records/search:
    parameters:
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/Contributor"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/InstrumentType"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/MsType"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/IonMode"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/Splash"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/CompoundName"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/CompoundClass"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/ExactMass"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/MassTolerance"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/Formula"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/Peaks"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/Intensity"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/NeutralLoss"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/PeakList"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/PeakListThreshold"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/InChI"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/InChIKey"
      - $ref: "./schemas/FilterParams.yaml#/components/parameters/Substructure"
    get:
      summary: Get a list of records as a search result in a specific format (accession, similarity score)
      operationId: getSearchResults
      responses:
        "200":
          description: "A list of search results."
          content:
            application/json:
              schema:
                $ref: "./schemas/SearchResult.yaml"
...